requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
Django==4.2.9
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.14.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
whitenoise==6.7.0
zope.interface==6.1
